Use LRO for createCDB/bootstrapCDB
Refactor database instance recconcilation as state machine based on
databaseInstanceReady.Reason.

tested=functional/integration tested
