+++ oracle/controllers/config_agent_helpers.go
@@ -16,6 +16,7 @@ package controllers
 
 import (
 	"context"
+	"fmt"
 
 	lropb "google.golang.org/genproto/googleapis/longrunning"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -33,7 +34,7 @@ func GetLROOperation(caClientFactory ConfigAgentClientFactory, ctx context.Conte
 }
 
 // DeleteLROOperation deletes LRO operation for the specified namespace instance and operation id.
-func DeleteLROOperation(caClientFactory ConfigAgentClientFactory, ctx context.Context, r client.Reader, namespace, id, instName string) error {
+func DeleteLROOperation(caClientFactory ConfigAgentClientFactory, ctx context.Context, r client.Reader, namespace, instName, id string) error {
 	caClient, closeConn, err := caClientFactory.New(ctx, r, namespace, instName)
 	if err != nil {
 		return err
@@ -42,4 +43,26 @@ func DeleteLROOperation(caClientFactory ConfigAgentClientFactory, ctx context.Co
 
 	_, err = caClient.DeleteOperation(ctx, &lropb.DeleteOperationRequest{Name: id})
 	return err
+}
+
+// Check for LRO job status
+// Return (true, nil) if LRO is done without errors.
+// Return (true, err) if LRO is done with an error.
+// Return (false, nil) if LRO still in progress.
+// Return (false, err) if other error occurred.
+func IsLROOperationDone(caClientFactory ConfigAgentClientFactory, ctx context.Context, r client.Reader, namespace, id, instName string) (bool, error) {
+	operation, err := GetLROOperation(caClientFactory, ctx, r, namespace, id, instName)
+	if err != nil {
+		return false, err
+	}
+	if !operation.GetDone() {
+		return false, nil
+	}
+
+	// handle case when remote LRO completed unsuccessfully
+	if operation.GetError() != nil {
+		return true, fmt.Errorf("Operation failed with err: %s. %v", operation.GetError().GetMessage(), err)
+	}
+
+	return true, nil
 }
