+++ oracle/pkg/k8s/condition.go
@@ -41,7 +41,9 @@ const (
 	CreateFailed                   = "CreateFailed"
 	CreateInProgress               = "CreateInProgress"
 	CreatePending                  = "CreatePending"
-	AwaitingRestore                = "AwaitingRestore"
+	BootstrapPending               = "BootstrapPending"
+	BootstrapInProgress            = "BootstrapInProgress"
+	RestorePending                 = "RestorePending"
 	ImportComplete                 = "ImportComplete"
 	ImportFailed                   = "ImportFailed"
 	ImportInProgress               = "ImportInProgress"
